+++ TestErrors.swift
@@ -1,0 +1,12 @@
+//
+//  TestErrors.swift
+//  RxSwiftExt
+//
+//  Created by Anton Nazarov on 24/05/2019.
+//  Copyright © 2019 RxSwiftCommunity. All rights reserved.
+//
+
+enum TestError: Error {
+    case dummyError
+}
+let testError = TestError.dummyError